requirements.txt
# OMtree Trading System Requirements
# Python 3.8+ recommended

# Core data processing
numpy>=1.21.0
pandas>=1.3.0

# Machine learning
scikit-learn>=1.0.0

# Visualization
matplotlib>=3.4.0
Pillow>=9.0.0

# Additional utilities (optional but recommended)
openpyxl>=3.0.9  # For Excel file support
yfinance>=0.2.18  # For downloading financial data
ta-lib==0.4.24  # Technical analysis library (if available for your platform)

# Performance monitoring (optional)
psutil>=5.9.0  # For system resource monitoring
tqdm>=4.65.0  # For progress bars